==> PythonCode/P1.py <==
#!conda install pip#!conda install secedgar#!conda install python-edgar#!pip install secedgar#!pip install python-edgar#from urllib.request import urlopen#from secedgar.filings import filings, FilingTypeimport edgarimport pandas as pdimport globimport csvimport reimport osimport numpy as npos.getcwd()os.chdir("/Users/lawrence/Documents/GitHub/DataScience/Sentimental_analysis_in_stock_trading")# Download 4-year edgar indexedgar.download_index("/Users/lawrence/Documents/GitHub/DataScience/Sentimental_analysis_in_stock_trading/Index", 2016, "fan.yuxin@outlook.com")# Input SP500 CIKos.chdir("/Users/lawrence/Documents/GitHub/DataScience/Sentimental_analysis_in_stock_trading")SP500 = pd.read_excel('SP500.xlsx', index_col=0)CIK = SP500['CIK']# Prepare table for web crawlingtsv_dir = "/Users/lawrence/Documents/GitHub/DataScience/Sentimental_analysis_in_stock_trading/Index"data_path = os.path.join(tsv_dir ,'*.tsv')indexlist = glob.glob (data_path)CIKlist = []Typelist = []Datelist = []Htmllist = []for iCIK in CIK:  print(iCIK)  for file in indexlist:     index = pd.read_csv(file, sep='\t', lineterminator='\n', names=None)     #test    #index = pd.read_csv('/Users/lawrence/Documents/GitHub/DataScience/Sentimental_analysis_in_stock_trading/Index/2018-QTR4.tsv', sep='\t', lineterminator='\n', names=None)    for item in index.iloc[:,0]:      if str(iCIK) in item and '10-Q' in item:        CIKlist.append(iCIK)        Typelist.append('10-Q')        Datelist.append(re.findall('10-Q\|([0-9].*?)\|edgar', item))        Htmllist.append(item)      if str(iCIK) in item and '10-K' in item:        CIKlist.append(iCIK)        Typelist.append('10-K')        Datelist.append(re.findall('10-K\|([0-9].*?)\|edgar', item))        Htmllist.append(item)   # Output tableos.chdir("/Users/yuxinfan/471Project")output = np.column_stack((CIKlist, Typelist, Datelist, Htmllist))writer=csv.writer(open('output.csv','w', newline=''))fields = ['CIKlist', 'Typelist', 'Datelist', 'Htmllist'] writer.writerow(fields)writer.writerows(output)
